Extract nest processing to separate module
